fix: database populator doesn't create applications properly
Due to the latest changes on the sources API's back end, the database
populator was erroring time and again. The main change on the sources
API has been that now sources can only have one compatible application
per application type, and therefore when the populator was trying to
POST applications which an application type that already existed on the
database, the back end complained.